rename: queries를 hooks 폴더 내로 이동 #11
