--- client/src/queries/queryClient.ts
@@ -1,12 +1,0 @@
-import { QueryClient } from '@tanstack/react-query';
-
-const queryClient = new QueryClient({
-  defaultOptions: {
-    queries: {
-      retry: 0,
-      suspense: true,
-    },
-  },
-});
-
-export default queryClient;
